Change data type of Model

+++ KSH.Api/Models/DTO/Request/KitDTO.cs
@@ -3,7 +3,7 @@
     public class KitDTO
     {
         public int Id { get; set; }
-        public int PurchaseCost { get; set; }
+        public long PurchaseCost { get; set; }
         public int Quantity { get; set; }
     }
 }
