+++ KSH.Api/Models/DTO/Request/LabDTO.cs
@@ -2,8 +2,8 @@
 {
     public class LabDTO
     {
-        public Guid Id { get; set; } 
-        public int PurchaseCost { get; set; } 
+        public Guid Id { get; set; }
+        public long PurchaseCost { get; set; }
         public int Quantity { get; set; }
     }
 }
