+++ KSH.Api/Models/DTO/Request/OrderCreateDTO.cs
@@ -18,17 +18,17 @@ namespace KSH.Api.Models.DTO.Request
 
         public string ShippingStatus { get; set; } = null!;
 
-        public string ShippingAddress { get; set; }= null!;
+        public string ShippingAddress { get; set; } = null!;
 
         public string PhoneNumber { get; set; } = null!;
 
         public bool IsLabDownloaded { get; set; }
 
-        public int Price { get; set; }
+        public long Price { get; set; }
 
-        public int Discount { get; set; }
+        public long Discount { get; set; }
 
-        public int TotalPrice { get; set; }
+        public long TotalPrice { get; set; }
 
         public string? Note { get; set; }
